temporarily remove sparks for now to launch without curated content
